+++ .gitignore
@@ -22,4 +22,3 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-
