--- src/components/Board.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Board() {
-  return (
-    <h1 style={{ margin: '0', color: 'black', textDecoration: 'none' }}>
-      보드
-    </h1>
-  );
-}
-
-export default Board;
