+++ src/components/Board.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Title = styled.h1`
+  margin: 0;
+  color: black;
+  text-decoration: none;
+`;
+
+function Board() {
+  return <Title>보드</Title>;
+}
+
+export default Board;
