--- src/components/BulletinBoard.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-function BulletinBoard() {
-  return (
-    <h1
-      style={{
-        margin: '0',
-        color: 'black',
-        textDecoration: 'none',
-        fontWeight: 'normal',
-      }}
-    >
-      게시판
-    </h1>
-  );
-}
-
-export default BulletinBoard;
