+++ src/components/BulletinBoard.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const Title = styled.h1`
+  margin: 0;
+  color: black;
+  text-decoration: none;
+  font-weight: normal;
+`;
+
+function BulletinBoard() {
+  return <Title>게시판</Title>;
+}
+
+export default BulletinBoard;
