+++ src/components/BulletinIcon.tsx
